moved specs
